refactor : Point를 MinePosition로 수정하고 일급컬렉션도 같이 수정
