--- src/main/kotlin/next/step/minesweeper/domain/mine/MinePoints.kt
@@ -1,6 +1,0 @@
-package next.step.minesweeper.domain.mine
-
-import next.step.minesweeper.domain.position.Position
-
-@JvmInline
-value class MinePoints(private val positions: Set<Position>) : Set<Position> by positions
