+++ src/main/kotlin/next/step/minesweeper/domain/mine/MinePositions.kt
@@ -1,0 +1,4 @@
+package next.step.minesweeper.domain.mine
+
+@JvmInline
+value class MinePositions(private val positions: Set<MinePosition>) : Set<MinePosition> by positions
