--- src/main/kotlin/next/step/minesweeper/domain/position/Position.kt
@@ -1,14 +1,0 @@
-package next.step.minesweeper.domain.position
-
-data class Position(val x: Int, val y: Int) {
-    init {
-        require(x >= BASE_X) { "위치 x는 $BASE_X 보다 작을 수 없습니다." }
-        require(y >= BASE_Y) { "위치 y는 $BASE_Y 보다 작을 수 없습니다." }
-    }
-
-    companion object {
-        private const val BASE_X = 0
-        private const val BASE_Y = 0
-        fun base() = Position(BASE_X, BASE_Y)
-    }
-}
